+++ CHANGELOG.md
@@ -8,4 +8,6 @@ The changes documented here do not include those from the original repository.
 
 ## [Unreleased]
 
+- Implements sendLocalNotification (that used to be setBadgeNumber)
+
 - Implements setBadge, getBadge and clearNotifications for Android (https://outsystemsrd.atlassian.net/browse/RMET-1561, https://outsystemsrd.atlassian.net/browse/RMET-1576, https://outsystemsrd.atlassian.net/browse/RMET-1557)